backend-engineer/problems/problem-2/requirements.txt
fastapi==0.111.0
uvicorn[standard]
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic-settings==2.3.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
pika==1.3.2
pytest==8.2.2
coverage==7.5.3